Move form validation to separate module
