+++ src/components/form-password-change/form-password-change.ts
@@ -4,45 +4,12 @@ import './form-password-change.css';
 
 import { Form, FormProps } from '../form/form';
 import { InputField } from '../input-field/input-field';
-import { Input } from '../../input';
+import { getPasswordRepeatGetError, passwordGetError } from '../../modules/validation';
 
 /**
  * Форма изменения пароля
  */
 export class FormPasswordChange extends Form {
-    /**
-     * Проверка валидности пароля
-     * @param {string} value - значение пароля
-     * @returns {string} - сообщение об ошибке или пустая строка, если ошибок нет
-     */
-    passwordGetError(value: string): string {
-        const isValid =
-            value.length >= 8 &&
-            /[a-z]+/.test(value) &&
-            /[A-Z]+/.test(value) &&
-            /[0-9]+/.test(value);
-        if (isValid) {
-            return '';
-        }
-        return 'Минимум 8 символов, содержит заглавные и строчные буквы и цифры';
-    }
-
-    /**
-     * Получить функцию для проверки повторного ввода пароля
-     * @param {Input} passwordInput - поле ввода пароля
-     * @returns {(value: string) => string} - функция для проверки повторного ввода пароля
-     */
-    getPasswordRepeatGetError(passwordInput: Input): (value: string) => string {
-        return (value: string) => {
-            const password = passwordInput.getValue();
-            const isValid = password === value;
-            if (!password || isValid) {
-                return '';
-            }
-            return 'Пароли не совпадают';
-        };
-    }
-
     /**
      * Обработчик нажатия на кнопку отправки формы
      */
@@ -71,10 +38,10 @@ export class FormPasswordChange extends Form {
             type: 'password',
             name: 'password',
             placeholder: 'Новый пароль',
-            getError: this.passwordGetError.bind(this),
+            getError: passwordGetError,
         });
 
-        const passwordRepeatGetError = this.getPasswordRepeatGetError(props.inputs.passwordInput);
+        const passwordRepeatGetError = getPasswordRepeatGetError(props.inputs.passwordInput);
         props.inputs.passwordRepeatInput = new InputField(root, {
             label: 'Повторите пароль',
             type: 'password',
