+++ src/components/form-profile-public/form-profile-public.ts
@@ -4,24 +4,12 @@ import './form-profile-public.css';
 
 import { Form, FormProps } from '../form/form';
 import { InputField } from '../input-field/input-field';
+import { organizationGetError } from '../../modules/validation';
 
 /**
  * Форма публичных данных профиля
  */
 export class FormProfilePublic extends Form {
-    /**
-     * Проверка валидности названия организации
-     * @param {string} value - значение названия организации
-     * @returns {string} - сообщение об ошибке или пустая строка, если ошибок нет
-     */
-    organizationGetError(value: string): string {
-        const isValid = value.trim().length >= 5;
-        if (isValid) {
-            return '';
-        }
-        return 'Минимум 5 символов';
-    }
-
     /**
      * Обработчик нажатия на кнопку отправки формы
      */
@@ -51,14 +39,13 @@ export class FormProfilePublic extends Form {
                 label: 'Название организации',
                 name: 'organization',
                 placeholder: 'Введите название',
-                getError: this.organizationGetError.bind(this),
+                getError: organizationGetError,
             }),
             descriptionInput: new InputField(root, {
                 type: 'text',
                 label: 'Описание',
                 name: 'description',
                 placeholder: 'Введите текст',
-                getError: this.organizationGetError.bind(this),
             }),
         };
